Оптимизация размера.
Размер кода уменьшен с 3900+ до 2048 байт занято RAM с 139 до 86. Есть шанс что заработает на ATTiny25.
Изменений очень много.
1. От Ардуины осталась только библиотека Wire. Убраны все millis mocros и прочее за ненадобностью.
2. Вся работа по прерываниям. Во время включения ESP сон PowerDown заменен на Idle.
3. Управление питанием реализовано отдельным классом.
4, Storage с кольцевым буфером не используется.
5. Обработка каналов счетчика сделана отдельным классом. В классе реализовано чтение и запись значения счетчиков в EEPROM
6. Функция подсчета контрольной суммы crc_8 заменена. Расчет в цикле. на ассемблере.
7. В классе SlaveI2C удален буфер.  Данные передаются непосредственно со структуры.
8. Подсчет времени работы. Отдельно считается минута (240 просыпаний) по истечении считается количество минут.  (Из кода убраны все умножение)
9 изменена структура данных. Счетчик, состояние и ацп сгруппированы в отдельную структуру.
10. изменен порядок байт в передаче времени пробуждения. передаются младший байт потом старший. В программе можно обратиться по адресу буфера как к 16битному значению.
11. Значение АЦП используется 8бит.

+++ ESP8266/src/master_i2c.h
@@ -38,8 +38,8 @@ struct SlaveData {
     uint8_t  state1;      //           вход 1
     uint32_t impulses0;   //Импульсов, канал 0
     uint32_t impulses1;   //           канал 1
-    uint16_t adc0;        //Уровень,   канал 0
-    uint16_t adc1;        //           канал 1
+    uint8_t adc0;        //Уровень,   канал 0
+    uint8_t adc1;        //           канал 1
     
     // HEADER_DATA_SIZE
